% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_bd.R
\name{bd_breaks}
\alias{bd_breaks}
\title{Date breaks corresponding to the first trading day of standard periods}
\usage{
bd_breaks(business.dates, n.max = 5)
}
\arguments{
\item{business.dates}{a vector of \code{Date} objects, sorted ascending}

\item{n.max}{the maximum number of breaks to return}
}
\value{
returns a function function: \code{max => [date range] => breaks} that generates the breaks 
        for the interval with the largest number of breaks less than \code{n.max}
}
\description{
The periods are:
\itemize{
 \item years
 \item quarters
 \item months
 \item weeks
 \item days
}
}

