% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bdscale.R
\docType{data}
\name{nyse}
\alias{nyse}
\title{Trading dates for the New York Stock Exchange extracted from the close prices of the S&P 500.}
\format{A vector of 16465 \code{Date} objects, starting on 1950-01-03 and ending on 2015-06-10}
\usage{
data(nyse)
}
\description{
Trading dates for the New York Stock Exchange extracted from the close prices of the S&P 500.
}

