% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scale_x_bd}
\alias{scale_x_bd}
\title{Position scale for a ggplot}
\usage{
scale_x_bd(..., expand = waiver(), breaks = bd_breaks(business.dates),
  minor_breaks = waiver(), business.dates)
}
\arguments{
\item{...}{see \code{\link{scale_x_date}}}

\item{expand}{see \code{\link{scale_x_date}}}

\item{breaks}{see \code{\link{scale_x_date}}}

\item{minor_breaks}{see \code{\link{scale_x_date}}}

\item{business.dates}{a vector of Date objects}
}
\description{
Position scale for a ggplot
}
\examples{
library(dplyr, quietly=T)
library(ggplot2)

nyse <- get_nyse()

set.seed(12345)
df <- data.frame(date=nyse, price=cumsum(rnorm(length(nyse))) + 100)

getwd()

calendar <- df \%>\% filter(as.Date('2014-08-01') < date & date < as.Date('2014-10-08')) \%>\% 
  ggplot(aes(x=date, y=price)) + geom_step()

ggsave(calendar, file='man/figures/calendar.PNG')

business <- df \%>\% filter(as.Date('2014-08-01') < date & date < as.Date('2014-10-08')) \%>\% 
  ggplot(aes(x=date, y=price)) + geom_step() + scale_x_bd(business.dates=nyse)

ggsave(business, file='man/figures/business.PNG')
}

