% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scale_x_bd}
\alias{scale_x_bd}
\title{Position scale for a ggplot}
\usage{
scale_x_bd(..., expand = waiver(), breaks = bd_breaks(business.dates),
  minor_breaks = waiver(), business.dates)
}
\arguments{
\item{...}{see \code{\link{scale_x_date}}}

\item{expand}{see \code{\link{scale_x_date}}}

\item{breaks}{see \code{\link{scale_x_date}}}

\item{minor_breaks}{see \code{\link{scale_x_date}}}

\item{business.dates}{a vector of Date objects}
}
\description{
Position scale for a ggplot
}
\examples{
library(magrittr)
library(dplyr)
library(ggplot2)

spx <- paste('http://real-chart.finance.yahoo.com/table.csv',
             's=\%5EGSPC&d=9&e=8&f=2014&g=d&a=0&b=3&c=1950&ignore=.csv',
             sep='?')
nyse <- spx \%>\% url \%>\% read.csv \%>\% extract2('Date') \%>\% as.Date

set.seed(12345)
df <- data.frame(date=nyse, price=cumsum(rnorm(length(nyse))) + 100)

df \%>\% filter(date > as.Date('2014-08-01')) \%>\% 
  ggplot(aes(x=date, y=price)) + geom_step()

df \%>\% filter(date > as.Date('2014-08-01')) \%>\% 
  ggplot(aes(x=date, y=price)) + geom_step() + scale_x_bd(business.dates=nyse)
}

