% Generated by roxygen2 (4.0.2): do not edit by hand
\name{yahoo}
\alias{yahoo}
\title{Get past trading days using close prices of supplied ticker}
\usage{
yahoo(ticker = "^GSPC")
}
\arguments{
\item{ticker}{The ticker you want to use, defaults to S&P 500: \code{^GSPC}}
}
\value{
a vector of Date
}
\description{
Get past trading days using close prices of supplied ticker
}

